Update ft_lstnew.spec.c
We weren't actually testing if ft_lstnew() was copying the address of content.
I corrected someone today who was using malloc() to copy the content of... content. I don't think this is the way to do it.
Example: if content is a pointer to a struct containing a malloced string, you can copy the struct but can only copy the address of the string, making it a partial copy.